requirements.txt
Pyrogram
TgCrypto
psycopg2-binary
ujson
python-dotenv
matplotlib
apscheduler==3.8.0
